pyproject.toml
[project]
name = "components-library"
version = "1.1.0"
description = "Python components library for FastHTML applications following atomic design principles"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "authlib>=1.3.0",
    "python-fasthtml>=0.12.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "mypy>=1.18.0",
    "ruff>=0.14.0",
    "pytest>=8.3.0",
    "pytest-cov>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["components_library"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true

[[tool.mypy.overrides]]
module = ["fasthtml.*", "authlib.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--verbose",
    "--strict-markers",
    "--tb=short",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
    "ARG",  # flake8-unused-arguments
    "SIM",  # flake8-simplify
    "N",    # pep8-naming
    "PL",   # pylint
    "RUF",  # ruff-specific rules
]

ignore = [
    "E501",    # line too long (handled by formatter)
    "PLR0913", # too many arguments
    "PLR0912", # too many branches
    "PLR2004", # magic value comparisons
    "RUF001",  # ambiguous unicode in strings (intentional for UI: ›, ℹ️, ×)
    "RUF002",  # ambiguous unicode in docstrings (intentional for UI examples)
    "PLC0415", # imports not at top-level (intentional lazy imports to avoid circular deps)
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # Allow unused imports in __init__.py

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[dependency-groups]
dev = [
    "mypy>=1.18.2",
    "pytest>=9.0.1",
    "ruff>=0.14.6",
]
